refactor: 1. noti -> notification 파일 이름 변경 2. useCalcElapsedTime -> calcCreatedToCurrenttime 변경 및 utils로 이동
