+++ src/interfaces/notification.ts
